Move tesla_common.js to a lib folder
Since you can't run `node tesla_common.js`, it seems like a good idea to move it to a separate folder.
