+++ lib/tesla_common.js
