+++ start_climate.js
@@ -10,7 +10,7 @@ const CLIMATE_ON_MINUTES = 10;
 /**** END USER CONFIG VARIABLES ****/
 
 
-const teslaCommon = require('./tesla_common');
+const teslaCommon = require('./lib/tesla_common');
 
 var vehicle;
